Main

Description

Este é um módulo definido para resolução dos tabuleiros propostos na TarefaD, relativa ao jogo de tabuleiro Sokoban, no âmbito da disciplina de Laboratórios de Informática I. Esta página servirá de documentação do código para tal tarefa.

Documentation

outStr :: [String] -> String

A função outStr é uma função pré-definida nas aulas para efectuar um unlines de forma mais eficaz.

main :: IO ()

tarefa4 :: [String] -> [String]

A função tarefa4 será responsável pela execução de todo o programa, com o intuito de calcular a solução. Este utilizará a função resolve, atribuindo-lhe dados como a lista de cardinais e a lista de caixas.

resolve :: [String] -> [((Int, Int), Int)] -> [(Int, Int)] -> [(Int, Int)] -> [Char] -> (Int, Int) -> Int -> String

A função resolve calculará a próxima posição através da averiguação do comando fornecido, assim como a existência de caixas e/ou paredes nos arredores da posição inicial. Esta função averiguará, através da função checkEnding, se todas as caixas se encontram nas posições dos pontos, devolvendo FIM nessa situação. Caso os comandos se esgotem, esta devolverá INCOMPLETO.

converte :: String -> (Int, Int)

A função converte servirá para posterior conversão da posição (o primeiro elemento de filtraCoord) para um par de inteiros, com respectivas coordenadas no eixo do x e no eixo do y.

>>> converte "2 3"
(2,3)

converteTudo :: [String] -> [(Int, Int)]

A função converteTudo efectuará o mesmo que a função converte, excepto que para listas de strings, de tal modo que devolverá uma lista de pares de inteiros.

filtraTab :: String -> String

A função filtraTab filtrará o tabuleiro no input inicial, e devolverá o mesmo input até ao ponto em que encontra um dígito.

filtraCoord :: String -> String

A função filtraCoord efectuará o inverso da função filtraTab; esta devolverá o input restante que teria sido filtrado em filtraTab.

checkChar :: [String] -> (Int, Int) -> Char -> Bool

A função checkChar receberá um tabuleiro e um par de coordenadas, assim como um caracter, para verificar a existência de esse caracter no tabuleiro.

>>> checkChar ["abcd"] (0,0) 'a'
True

checkCardList :: [String] -> (Int, Int) -> [(Int, Int)]

A função checkCardList efectuará uma simples operação: a compilação sob forma de lista de todos os cardinais existentes no tabuleiro, para uma posterior comparação.

checkDotList :: [String] -> (Int, Int) -> [(Int, Int)]

A função checkCardList assemelha-se à função checkCardList, diferindo apenas no caracter em questão (descobre todos os pontos).

boxIdentify :: [(Int, Int)] -> Int -> [((Int, Int), Int)]

A função boxIdentify atribui um índice a cada uma das caixas existentes, de forma a poder controlá-las via funções auxiliares, para que sejam alteradas consoante a sua movimentação.

boxSearch :: [((Int, Int), Int)] -> (Int, Int) -> Int

A função boxSearch está encarregue de comparar a caixa que está prestes a ser movimentada, para que possa ser alterada na lista de caixas existentes.

boxModify :: [((Int, Int), Int)] -> Int -> (Int, Int) -> [((Int, Int), Int)]

A função boxModify modifica as caixas consoante a posição destas se altera no tabuleiro. Servirá de base para comparar com a lista de pontos.

checkEnding :: [((Int, Int), Int)] -> [(Int, Int)] -> Bool

A função checkEnding verifica se o fim do jogo terá sido atingido, consoante a igualdade entre a lista de caixas existentes e a lista de pontos presentes no tabuleiro.