Main

Description

Este é um módulo definido para resolução dos tabuleiros propostos na Tarefa1, relativa ao jogo de tabuleiro Sokoban, no âmbito da disciplina de Laboratórios de Informática I. Esta página servirá de documentação do código para tal tarefa.

Documentation

outStr :: [String] -> String

A função outStr é uma função pré-definida nas aulas para efectuar um unlines de forma mais eficaz.

main :: IO ()

tarefa1 :: [String] -> [String]

A função tarefa1 será responsável por receber o input sob forma de lista de strings, no qual estarão incluídos o tabuleiro, posição e coordenadas de caixas.

verificaTudo :: String -> [String]

A função verificaTudo é o cérebro de todo este programa. Esta função receberá o mesmo tabuleiro que na função tarefa1, apenas com a diferença de estar sob forma de string, e, através de múltiplas operações via funções auxiliares, verificará todos os casos consoante a validade dos mesmos.

filtraTab :: String -> String

A função filtraTab filtrará o tabuleiro no input inicial, e devolverá o mesmo input até ao ponto em que encontra um dígito.

filtraCoord :: String -> String

A função filtraCoord efectuará o inverso da função filtraTab; esta devolverá o input restante que teria sido filtrado em filtraTab.

converte :: String -> (Int, Int)

A função converte servirá para posterior conversão da posição (o primeiro elemento de filtraCoord) para um par de inteiros, com respectivas coordenadas no eixo do x e no eixo do y.

>>> converte "2 3"
(2,3)

converteTudo :: [String] -> [(Int, Int)]

A função converteTudo efectuará o mesmo que a função converte, excepto que para listas de strings, de tal modo que devolverá uma lista de pares de inteiros.

checkChar :: [String] -> (Int, Int) -> Char -> Bool

A função checkChar receberá um tabuleiro e um par de coordenadas, assim como um caracter, para verificar a existência de esse caracter no tabuleiro.

>>> checkChar ["abcd"] (0,0) 'a'
True

validaDimensao :: [String] -> Int -> Int

A função validaDimensao apenas verificará se as dimensões do tabuleiro são válidas, a partir da comparação do tamanho de cada uma das linhas. Caso não seja, devolverá o número da linha cujo comprimento diferente da anterior.

validaPosicao :: [String] -> [String] -> Int

A função validaPosicao averiguará se a posição recebida está dentro das dimensões estipuladas do tabuleiro, assim como analisará se esta não se encontra em cima de uma parede (#).

checkSpot :: [String] -> (Int, Int) -> Bool

A função checkSpot funcionará em conjunto com uma outra função, de modo a confirmar se o tabuleiro é apenas composto por cardinais, pontos e espaços. Caso não seja, devolve False.

checkTab :: [String] -> (Int, Int) -> Int

Através da função checkSpot, a função checkTab partirá através do valor obtido pela função checkSpot para averiguar se o caracter em que se encontra é um caracter válido. Caso este não seja, a função devolverá a linha em que essa condição não se verifique

checkCardList :: [String] -> (Int, Int) -> [(Int, Int)]

A função checkCardList efectuará uma simples operação: a compilação sob forma de lista de todos os cardinais existentes no tabuleiro, para uma posterior comparação.

checkDotList :: [String] -> (Int, Int) -> [(Int, Int)]

A função checkDotList funcionará do mesmo modo que a função checkCardList, com a diferença de compilar uma lista de todos os pontos existentes no tabuleiro.}

compareBoxDots :: [String] -> [(Int, Int)] -> Int

A função compareBoxDots servirá para comparação entre o número de caixas e o número de pontos. Como, para que um tabuleiro seja válido, o número de caixas terá de ser igual ao número de pontos existentes, esta função, através da lista compilada pela função checkDotList, comparará o número de caixas com o número de pontos.

noRepetitions :: [String] -> Int -> Int

A função noRepetitions verificará se não há uma repetição de caixas, isto é, não poderão estar duas caixas na mesma posição. Caso existam duas caixas sob essa condição, será devolvida a linha da caixa que se está a tentar sobrepor.

outOfBounds :: [String] -> [(Int, Int)] -> Int -> Int

A função outOfBounds apenas comparará a posição das caixas com as dimensões do tabuleiro, isto é, as caixas não podem estar fora das dimensões do tabuleiro. Devolve a linha em que tal não se verifique.